Module Mirage_crypto_rng

Randomness

Usage notes

type bits = int

Interface

type g

A generator (PRNG) with its state.

exception Unseeded_generator

Thrown when using an uninitialized generator.

module type Generator = sig ... end

A single PRNG algorithm.

type 'a generator = (module Generator with type g = 'a)

module Fortuna : Generator

Fortuna, a CSPRNG proposed by Schneier.

module Hmac_drbg : sig ... end

HMAC_DRBG: A NIST-specified RNG based on HMAC construction over the provided hash.

module Null : Generator

No-op generator returning exactly the bytes it was seeded with.

val create : ?⁠g:'a -> ?⁠seed:Cstruct.t -> ?⁠strict:bool -> (module Generator with type g = 'a) -> g

create module uses a module conforming to the Generator signature to instantiate the generic generator g.

g is the state to use, otherwise a fresh one is created.

seed can be provided to immediately reseed the generator with.

strict puts the generator into a more standards-conformant, but slighty slower mode. Useful if the outputs need to match published test-vectors.

val generator : g Stdlib.ref

Default generator. Functions in this module use this generator when not explicitly supplied one.

Swapping the generator is a way to subvert the random-generation process e.g. to make it fully deterministic. Don't do that unless you know what you're doing, but use Mirage_crypto_entropy.initialize or Mirage_crypto_rng_unix.initialize ().

generator defaults to Null.

val generate : ?⁠g:g -> int -> Cstruct.t

Invoke generate on g or default generator.

val block : g option -> int

Block size of g or default generator.

Examples

Generating a random 13-byte Cstruct.t:

let cs = Rng.generate 13

Generating a list of Cstruct.t, passing down an optional generator:

let rec f1 ?g ~n i =
if i < 1 then [] else Rng.generate ?g n :: f1 ?g ~n (i - 1)

Generating a Z.t smaller than 10:

let f2 ?g () = Mirage_crypto_pk.Z_extra.gen ?g Z.(~$10)

Creating a local Fortuna instance and using it as a key-derivation function:

let f3 secret =
let g = Rng.(create ~seed:secret (module Generators.Fortuna)) in
Rng.generate ~g 32